requirements.txt
amqp==5.0.3
appdirs==1.4.4
asgiref==3.2.7
attrs==20.3.0
billiard==3.6.3.0
boto3==1.16.47
botocore==1.19.47
celery==5.0.5
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.0.6
django-admin-jqueryui==1.8.22
django-admin-rangefilter==0.6.3
django-colorfield==0.3.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-grappelli==2.14.3
django-inline-ordering==1.0.2
django-modeltranslation==0.16.1
django-redis==4.12.1
django-rest-passwordreset==1.1.0
django-richtextfield==1.6
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
drf-yasg==1.20.0
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.11
identify==1.5.8
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
modeltranslation-grappelli==3.0.0
nodeenv==1.5.0
packaging==20.8
Pillow==8.0.1
pre-commit==2.8.2
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
PyJWT==2.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
raven==6.10.0
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sentry-sdk==1.1.0
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.1.0
wcwidth==0.2.5
